single endpoint is not single.ejs
