+++ services/fleet.js
@@ -62,6 +62,41 @@ async function buildStatistics(){
     return data;
 };
 
+async function getScriptBySingleEndpoint(endpoint) {
+    // keeping all scripts in a variable
+    let allScripts = [];
+    
+    // building the API url with the correct endpoint id each time
+    let scriptUri = `/api/v1/fleet/hosts/${endpoint.id}/scripts`;
+    try {
+        // sending the request and listening for response
+        const response = await fleetApiGetRequest(scriptUri);
+        const scripts = response.scripts;
+        if (scripts) {
+            
+            scriptsWithExecutionDetails = await fillScriptExecutionDetails(scripts);
+            allScripts = allScripts.concat(scriptsWithExecutionDetails.map(script => (
+                {
+                endpoint_id: endpoint.id,
+                endpoint: endpoint.hostname,
+                script: script.name,
+                execution_time: script.last_execution ? script.last_execution.executed_at : 'N/A',
+                status: script.last_execution ? script.last_execution.status : 'N/A',
+                execution_id: script.execution_id,
+                message: script.message,
+                output: script.output ? Buffer.from(script.output).toString('base64') : null,
+                exit_code: script.exit_code,
+                script_contents: script.script_contents ? Buffer.from(script.script_contents).toString('base64') : null,
+            })));
+        } else {
+            console.warn(`No scripts found for endpoint ${endpoint.id}`);
+        }
+    } catch (error) {
+        console.error(`Error fetching scripts for endpoint ${endpoint.id}:`, error);
+    }
+    return allScripts;
+}
+
 async function getScriptByEndpoint(endpointList) {
     // keeping all scripts in a variable
     let allScripts = [];
@@ -285,6 +320,7 @@ module.exports = {
     getEndpoint,
     buildStatistics,
     getAgentEnrollCmd,
+    getScriptBySingleEndpoint,
     getScriptByEndpoint,
     mergeEndpointAndScripts, 
     fleetApiPostRequest,
