+++ views/single_endpoint.ejs
@@ -1,0 +1,106 @@
+<%- include('partials/header') %>
+<link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css">
+<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"></script>
+<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
+<link rel="stylesheet" href="/css/endpoints.css">
+
+<div class="container">
+    <!-- Endpoint cards -->
+    <div class="row">
+        <div class=”cust-grid-col”>
+        <div class="card fixed-sized-card">
+            <div class="card-body">
+                <% 
+                let platformLogo;
+                switch (endpoint.platform) {
+                    case "rhel":
+                        platformLogo = "/img/rhel_logo.png";
+                        break;
+                    case "kali":
+                        platformLogo = "/img/kali_logo.png";
+                        break;
+                    case "ubuntu":
+                        platformLogo = "/img/ubuntu_logo.png";
+                        break;
+                    case "debian":
+                        platformLogo = "/img/debian_logo.png";
+                        break;
+                    case "windows":
+                        platformLogo = "/img/windows_logo.png";
+                        break;
+                    case "centos":
+                        platformLogo = "/img/centos_logo.png";
+                        break;
+                    case "macos":
+                        platformLogo = "/img/macos_logo.png";
+                        break;
+                    default:
+                        platformLogo = "/img/generic_logo.png";
+                        break;
+                }
+                %>
+                <a href="/endpoints/<%= endpoint.id %>"><img src="<%= platformLogo %>" class="card-img-top platform-img"></a>
+                <h5 id="endpoint_hostname" class="card-title">
+                    <%= endpoint.hostname %>
+                </h5>
+                <p class="card-text">
+                    <strong>IP:</strong> <%= endpoint.primary_ip %>
+                </p>
+                <p class="card-text">
+                    <strong>MAC:</strong> <%= endpoint.primary_mac %>
+                </p>
+                <p class="card-text">
+                    <strong>OS:</strong> <%= endpoint.os_version %>
+                </p>
+                <p class="card-text">
+                    <strong>Uptime:</strong> <%= (endpoint.uptime / (1000 * 60 * 60 * 24)).toFixed(2) %> days
+                </p>
+                <p class="card-text">
+                    <strong>Last Seen:</strong> <%= endpoint.formatted_last_seen %>
+                </p>
+                <p class="card-text">
+                    <strong>Last Scan:</strong>
+                    <% if (endpoint.formatted_last_scan && endpoint.formatted_last_scan !== '1/1/1970, 00:00:00') { %>
+                        <%= endpoint.formatted_last_scan %>
+                    <% } else { %>
+                        N/A
+                    <% } %>
+                </p>
+                <p id="endpoint_id" hidden class="endpointID">
+                    <%- endpoint.id %>
+                </p>
+                <!-- Dropdown menu -->
+                <div class="dropdown position-absolute" style="top: 10px; right: 10px;">
+                    <button class="btn dropdown-toggle custom-dropdown" type="button" id="dropdownMenuButton<%= endpoint.uuid %>" data-bs-toggle="dropdown" aria-expanded="false">
+                        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-three-dots-vertical" viewBox="0 0 16 16">
+                            <path d="M9.5 13a1.5 1.5 0 1 1-3 0 1.5 1.5 0 0 1 3 0m0-5a1.5 1.5 0 1 1-3 0 1.5 1.5 0 0 1 3 0m0-5a1.5 1.5 0 1 1-3 0 1.5 1.5 0 0 1 3 0"/>
+                        </svg>
+                    </button>
+                    <ul class="dropdown-menu" aria-labelledby="dropdownMenuButton<%= endpoint.uuid %>">
+                        <% if (endpoint.chipsec) { %>
+                            <li><a id="execute_script" class="dropdown-item" href="#">Execute a script</a></li>
+                            <li><a id="run_module" class="dropdown-item" href="/endpoints/<%= endpoint.id %>">Run a chipsec module</a></li>
+                            <li><a id="uninstall_chipsec" class="dropdown-item" href="#">Uninstall chipsec</a></li>
+                            <li><a id="delete_host" class="dropdown-item" href="#">Delete host</a></li>
+                        <% } else { %>
+                            <li><a id="install_chipsec" class="dropdown-item" href="#">Install chipsec</a></li>
+                            <li><a id="delete_host" class="dropdown-item" href="#">Delete host</a></li>
+                        <% } %>
+                    </ul>
+                </div>
+            </div>
+            <div class="card-footer">
+                <% if (endpoint.chipsec) { %>
+                    <div class="text-muted chipsec-status">✅ Chipsec installed</div>
+                <% } else { %>
+                    <div class="text-muted chipsec-status">⛔ Chipsec isn't installed </div>
+                <% } %>
+            </div>
+        </div>
+    </div>
+    </div>
+    <%- include('partials/chipsec_modules') %>
+</div>
+<script src="/js/single_endpoint.js"></script>
+<%- include('partials/footer') %>
+</div>
